Changed infrastructure of game
Instead of an array of ints, it's now an array of dicts. two keys: "value" and "ID".
Value holds the integer, ID is only used to prevent the 'even number exploit' which involves splitting an even number then canceling it with itself. The ID is generated after each number is made or interacted with. Top row nums take on the ID of their source bottom row number, and nums with matching ID's can't cancel.

Last thing to do is really just to comment the code better and clean it up, as i'm sure there's a bunch of weird stuff gong on in here.